requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django>=4.2.0
djangorestframework>=3.14.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0